Creating the Fetch and Load component classes.
